pmemcheck: tests: Move amd64 specific tests to separate directory
The subsequent patches add ppc64 support for the pmemcheck.
Prepare to have independent test results for consistency.
