+++ pmemcheck/tests/amd64/filter_stderr
@@ -1,0 +1,4 @@
+#! /bin/sh
+
+# Use the generic Pmemcheck stderr filter
+../filter_stderr
