+++ pmemcheck/tests/amd64/floats.stderr.exp
