+++ pmemcheck/tests/amd64/floats.vgtest
