+++ pmemcheck/tests/amd64/nt_stores.vgtest
