+++ pmemcheck/tests/amd64/sse_stores.stderr.exp
