+++ pmemcheck/tests/amd64/sse_stores.vgtest
