+++ pmemcheck/tests/logging_related/amd64/filter_stderr
@@ -1,0 +1,4 @@
+#! /bin/sh
+
+# Use the generic Pmemcheck logging_related test's stderr filter
+../filter_stderr
