+++ pmemcheck/tests/logging_related/amd64/logging.stderr.exp
